asuc.org: add ftp subdomain
"The site is pantry.asuc.org. I was told the host name is ftp.asuc.org
while the user name is pantry.asuc.org. It used to work, so not being
able to log-in is a new development."

I guess we forgot to add this when migrating DNS? Probably because it
wasn't listed in the subdomain list and just added implicitly by evil
cPanel.