Add back forgotten hostnames
We lost a few hostnames in the past few years, so add
them back to the unused list. Of particular note are
apocalypse and conquest, which were the twin login servers
before tsunami, and part of a long-lived reference to the
Four Horsemen of the Apocalypse - the others being
death, famine, pestilence, and war, which still exist.

In 2011 we also considered adding the following hostnames:
blackice
budgetcut
bushfire
cavein
cloudburst
collision
deathbymonkeys
doom
lotso
oilspill
shipwreck
spike
strike
thunderstorm
walkout